A new take on the LBANN toolkit.
This line of development will focus on providing optimizations for
machine learning applications in PyTorch with a particular focus on
HPC systems. The initial capability introduces the beginnings of an
LBANN "backend" to PyTorch, primarily implementing memory
optimizations for the MI300A architecture of El Capitan.

tl;dr: initial commit.